[ROCm/gfx1250] Fix DeepSeek-R1 MXFP4 accuracy: fp32 softmax-p in MLA attention P.V dot; a8w4 MoE weight shuffle
On gfx1250 the bf16 matmul path is lossy (~0.16-0.28% vs fp64; fp32 is ~1e-6),
whereas gfx950 (MFMA) tolerates it. The triton MLA attention downcast the softmax
weights `p` to bf16 before the P.V `tl.dot`, so that per-op error accumulated over
61 layers into a large GSM8K drop (0.811 vs gfx950 0.932). Keep `p` in fp32:

- triton_ops/decode_attention.py (_fwd_grouped_kernel_stage1): P.V dot
- triton_ops/extend_attention.py (_fwd_kernel prefix + extend-local): P.V dot
  tl.dot(p, v.to(tl.float32), out_dtype=tl.float32)  # no p.to(v.dtype)

QK needs no change (tl.dot(bf16,bf16) already accumulates fp32). Result: real
a8w4 MoE + cuda-graph GSM8K 1319Q = 0.950 (was 0.811), full speed ~155 tok/s.

Also enable the gfx1250 a8w4 MoE weight shuffle in quark_w4a4_mxfp4_moe.py
(SGLANG_MOE_SHUFFLE_GFX1250, default on when AITER_FORCE_A8W4): the grouped a8w4
kernel consumes (16,16)-preshuffled FP4 weights and the B-scale is already
n32k4-shuffled for gfx1250, so an unshuffled weight mismatches it -> garbage.

run_ds-r1.sh: model path + SGLANG_MOE_SHUFFLE_GFX1250=1.

+++ python/sglang/kernels/ops/attention/decode_attention.py
@@ -524,7 +524,10 @@ def _fwd_grouped_kernel_stage1(
             re_scale = tl.exp(e_max - n_e_max)
             p = tl.exp(qk - n_e_max[:, None])
             acc *= re_scale[:, None]
-            acc += tl.dot(p.to(v.dtype), v)
+            # Keep the softmax weights p in fp32 for the P·V dot (do NOT downcast p to
+            # bf16). The bf16 downcast of p was the accuracy loss vs a torch fp32 SDPA
+            # reference (recovers gfx1250 R1 GSM8K ~0.82 -> ~0.92 with attention idealized).
+            acc += tl.dot(p, v.to(tl.float32), out_dtype=tl.float32)
 
             e_sum = e_sum * re_scale + tl.sum(p, 1)
             e_max = n_e_max
